Fix several problems with install (all OS)
In check and deploy phases, we were not actually installing the code
There is also a change in the naming convention, putting GCC version
before the MPI version. This is to stick with the convention in other
projects.

Install is done in all OS.

Updated the modulefile generation to set CFLAGS as a variable
instead of a path.

Added modulefile checks in all OS.